HomeCar Loan/pythonfsdgroup4capstone/dj-be/profiles-rest-api/requirements.txt
django
Djangorestframework
pymysql